# switch to premultiplied alpha, faster blurs