… src/pixie/masks.nim
… src/pixie/paths.nim
… tests/benchmark_blends.nim
… tests/benchmark_masks.nim
… tests/images/masks/shifted.png
… tests/images/masks/spread.png
… tests/test_masks.nim