```nim
let image = newImage(2560, 1440)

timeIt "blur":
  image.blur(40)

before:
blur ............................ 3458.205 ms   3490.179 ms   ±45.217     x2
after:
blur ............................ 1572.042 ms   1576.937 ms    ±8.472     x4
```

```nim
let mask = newMask(2560, 1440)

timeIt "blur":
  mask.blur(40)

before:
blur ............................ 1057.261 ms   1067.605 ms   ±10.988     x5
after:
blur ............................. 495.492 ms    499.507 ms    ±5.113    x10
```