@guzba 2021-02-10
before:
`blendExclusion ..................... 0.364 ms      0.369 ms    ±0.016  x1000`
after:
`blendExclusion ..................... 0.230 ms      0.232 ms    ±0.002  x1000`